backend/requirements.txt
# Core Web Frameworks (FastAPI)
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
asyncpg>=0.29.0

# Quant & Data Science
yfinance>=0.2.36
numpy>=1.24.0
pandas>=2.0.0
PyYAML>=6.0.1

# Database & Infrastructure
psycopg2-binary>=2.9.0
apscheduler>=3.10.0
redis>=5.0.0
psutil>=5.9.0
python-dotenv>=1.0.0

# Security & Validation
pyjwt>=2.8.0
marshmallow>=3.20.0

# Observability & Metrics
prometheus-client>=0.20.0

# Testing
pytest>=7.4.0
